requirements.txt
Pillow
celery
requests
django-celery-beat
pytz
redis
djangorestframework
openai